プログラムを2つに分割した
