+++ MazeCreate/MazeCreate.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 16.0.29503.13
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "MazeCreate", "MazeCreate\MazeCreate.vcxproj", "{D6045D37-C71E-494C-8857-429C7CA1127A}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "MazeCreate2", "..\..\1年後期 C++Test\r_wall\MazeCreate2\MazeCreate2.vcxproj", "{8CF7E239-5401-4EC4-ACAF-EE5CA29C81DC}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "MazeCreate2", "MazeCreate2\MazeCreate2.vcxproj", "{4959EF09-4C3E-4945-8A0F-F4C44B22479B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,14 +23,14 @@ Global
 		{D6045D37-C71E-494C-8857-429C7CA1127A}.Release|x64.Build.0 = Release|x64
 		{D6045D37-C71E-494C-8857-429C7CA1127A}.Release|x86.ActiveCfg = Release|Win32
 		{D6045D37-C71E-494C-8857-429C7CA1127A}.Release|x86.Build.0 = Release|Win32
-		{8CF7E239-5401-4EC4-ACAF-EE5CA29C81DC}.Debug|x64.ActiveCfg = Debug|x64
-		{8CF7E239-5401-4EC4-ACAF-EE5CA29C81DC}.Debug|x64.Build.0 = Debug|x64
-		{8CF7E239-5401-4EC4-ACAF-EE5CA29C81DC}.Debug|x86.ActiveCfg = Debug|Win32
-		{8CF7E239-5401-4EC4-ACAF-EE5CA29C81DC}.Debug|x86.Build.0 = Debug|Win32
-		{8CF7E239-5401-4EC4-ACAF-EE5CA29C81DC}.Release|x64.ActiveCfg = Release|x64
-		{8CF7E239-5401-4EC4-ACAF-EE5CA29C81DC}.Release|x64.Build.0 = Release|x64
-		{8CF7E239-5401-4EC4-ACAF-EE5CA29C81DC}.Release|x86.ActiveCfg = Release|Win32
-		{8CF7E239-5401-4EC4-ACAF-EE5CA29C81DC}.Release|x86.Build.0 = Release|Win32
+		{4959EF09-4C3E-4945-8A0F-F4C44B22479B}.Debug|x64.ActiveCfg = Debug|x64
+		{4959EF09-4C3E-4945-8A0F-F4C44B22479B}.Debug|x64.Build.0 = Debug|x64
+		{4959EF09-4C3E-4945-8A0F-F4C44B22479B}.Debug|x86.ActiveCfg = Debug|Win32
+		{4959EF09-4C3E-4945-8A0F-F4C44B22479B}.Debug|x86.Build.0 = Debug|Win32
+		{4959EF09-4C3E-4945-8A0F-F4C44B22479B}.Release|x64.ActiveCfg = Release|x64
+		{4959EF09-4C3E-4945-8A0F-F4C44B22479B}.Release|x64.Build.0 = Release|x64
+		{4959EF09-4C3E-4945-8A0F-F4C44B22479B}.Release|x86.ActiveCfg = Release|Win32
+		{4959EF09-4C3E-4945-8A0F-F4C44B22479B}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
