+++ MazeCreate/MazeCreate/MazeCreate.vcxproj
@@ -123,6 +123,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="main.cpp" />
     <ClCompile Include="MazeCreate.cpp" />
     <ClCompile Include="RouteSearch.cpp" />
   </ItemGroup>
