+++ MazeCreate/MazeCreate/MazeCreate.vcxproj.filters
@@ -21,6 +21,9 @@
     <ClCompile Include="RouteSearch.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
+    <ClCompile Include="main.cpp">
+      <Filter>ソース ファイル</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="MazeCreate.h">
